# docs/swagger.yaml
basePath: /api
definitions:
  model.BookingResponse:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      room:
        $ref: '#/definitions/model.Room'
      room_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/model.BookingStatus'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: string
    type: object
  model.BookingStatus:
    enum:
    - active
    - cancelled
    type: string
    x-enum-varnames:
    - BookingStatusActive
    - BookingStatusCancelled
  model.CreateBookingInput:
    properties:
      end_time:
        type: string
      room_id:
        type: string
      start_time:
        type: string
      user_id:
        type: string
    required:
    - end_time
    - room_id
    - start_time
    - user_id
    type: object
...
  model.LoginRequest:
    properties:
      email:
        example: riparuk@gmail.com
        type: string
      password:
        example: strongpassword
        type: string
    required:
    - email
    - password
    type: object
  model.RegisterRequest:
    properties:
      email:
        example: riparuk@gmail.com
        type: string
      name:
        example: Rifa Faruqi
        type: string
      password:
        example: strongpassword
        type: string
    required:
    - email
    - name
    - password
    type: object
  model.Room:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.UpdateBookingInput:
    properties:
      end_time:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/model.BookingStatus'
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
    type: object
  model.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API documentation for Meet Book API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Meet Book API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Email
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: Email
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Register
      tags:
      - auth
  /bookings:
    post:
      consumes:
      - application/json
      description: Create a new room booking
      parameters:
      - description: Booking details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateBookingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BookingResponse'
      security:
      - BearerAuth: []
      summary: Create a new booking
      tags:
      - bookings
  /bookings/{id}:
    get:
      description: Get a booking by its ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BookingResponse'
      security:
      - BearerAuth: []
      summary: Get a booking by ID
      tags:
      - bookings
    put:
      consumes:
      - application/json
      description: Update an existing booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Booking update details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateBookingInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BookingResponse'
      security:
      - BearerAuth: []
      summary: Update a booking
      tags:
      - bookings
  /bookings/{id}/cancel:
    post:
      description: Cancel an existing booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BookingResponse'
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - bookings
  /bookings/room/{room_id}:
    get:
      description: Get a list of all bookings for a specific room
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.BookingResponse'
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: Get bookings for a specific room
      tags:
      - bookings
  /bookings/room/{room_id}/{date}:
    get:
      description: Get a list of all bookings for a specific room on a specific date
        with optional status filter
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: string
      - description: 'Date (format: YYYY-MM-DD)'
        in: path
        name: date
        required: true
        type: string
      - description: Filter by status (e.g., 'active', 'cancelled')
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of bookings
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.BookingResponse'
                type: array
            type: object
        "400":
          description: Invalid room ID or date format
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Failed to fetch room bookings
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get bookings for a specific room on a specific date
      tags:
      - bookings
  /bookings/upcoming:
    get:
      description: Get a list of all upcoming bookings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BookingResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get upcoming bookings
      tags:
      - bookings
  /bookings/users/{user_id}:
    get:
      description: Get all bookings for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BookingResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get all bookings for a user
      tags:
      - bookings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"